refactor: move useAuth to separate file
