+++ src/components/account-details.tsx
@@ -5,8 +5,8 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "./ui/button";
-import { useAuth } from "./auth-provider";
 import { LogOut } from "lucide-react";
+import { useAuth } from "@/hooks/useAuth";
 
 export function AccountDetails() {
   const { user, signOut } = useAuth();
