+++ src/hooks/useAuth.tsx
@@ -1,0 +1,6 @@
+import { AuthContext } from "@/components/auth-provider";
+import { useContext } from "react";
+
+export function useAuth() {
+  return useContext(AuthContext);
+}
